global update routes

+++ routes/users.js
@@ -25,9 +25,13 @@ router.post("/signup", (req, res) => {
   // Check if the user has not already been registered
   User.findOne({ email: req.body.email }).then((data) => {
     const hash = bcrypt.hashSync(req.body.password, 10);
-    const patternEmail = /^[a-zA-Z0-9._%+-]{1,64}@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
-    if (!patternEmail.test(req.body.email)){
-      return res.json({ result: false, error: "Le format de l'email n'est pas valide" })
+    const patternEmail =
+      /^[a-zA-Z0-9._%+-]{1,64}@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
+    if (!patternEmail.test(req.body.email)) {
+      return res.json({
+        result: false,
+        error: "Le format de l'email n'est pas valide",
+      });
     }
     if (data === null) {
       const token = uid2(32);
@@ -47,30 +51,6 @@ router.post("/signup", (req, res) => {
           token,
           email: req.body.email,
         });
-      });
-    } else {
-      // User already exists in database
-      res.json({ result: false, error: "Cet email est déjà utilisé" });
-    }
-  });
-});
-//Route POST pour l'inscription user via google
-router.post("/google", (req, res) => {
-  // Check if the user has not already been registered
-  User.findOne({ email: req.body.username }).then((data) => {
-    const hash = bcrypt.hashSync(req.body.password, 10);
-
-    if (data === null) {
-      const newUser = new User({
-        firstname: req.body.fisrtname,
-        lastname: req.body.lastname,
-        reservations: [null],
-        token,
-      });
-
-      newUser.save().then(() => {
-        res.json({ result: true, firstname: data.firstname, token });
-        console.log(token);
       });
     } else {
       // User already exists in database
@@ -100,40 +80,6 @@ router.post("/signin", (req, res) => {
       res.json({ result: false, error: "Aucun utilisateur trouvé" });
     }
   });
-});
-
-// Route pour la connexion avec Google
-router.post("/google", async (req, res) => {
-  const { idToken } = req.body; // Le token reçu du client
-
-  try {
-    // Vérifier le token Google
-    const ticket = await client.verifyIdToken({
-      idToken,
-      audience: process.env.GOOGLE_CLIENT_ID, // ID client Google
-    });
-    const payload = ticket.getPayload(); // Récupérer les informations de l'utilisateur
-
-    // Vérifier si l'utilisateur existe déjà dans la base de données
-    let user = await findOne({ googleId: payload.sub });
-    if (!user) {
-      // Créer un nouvel utilisateur si non existant
-      user = new user({
-        email: payload.email,
-        googleId: payload.sub,
-        name: payload.name,
-      });
-      await user.save(); // Sauvegarder le nouvel utilisateur
-    }
-
-    // Gérer la session ou générer un token (JWT, par exemple)
-    const token = user.generateAuthToken(); // Méthode pour générer un token
-    res.status(200).json({ message: "Connexion réussie.", token });
-  } catch (error) {
-    res
-      .status(500)
-      .json({ message: "Erreur du serveur.", error: error.message });
-  }
 });
 
 //route qui permet de récupérer les groupes à l'aide du token de l'utilisateur
@@ -174,23 +120,25 @@ router.put("/createGroup/:token", (req, res) => {
 
 //route pour rejoindre un group via son email
 router.put("/joinGroup", (req, res) => {
-  if (!checkBody(req.body, ["email" , "id"])) {
+  if (!checkBody(req.body, ["email", "id"])) {
     res.json({ result: false, error: "Champs manquants à remplir" });
     return;
   }
-  User.findOne({ email: req.body.email , groups :{$nin: req.body.id}}).then((data) => {
-    if (!data) {
-      res.json({
-        result: false,
-        error: "Utilisateur dans le groupe ou inexistant",
-      });
-      return;
+  User.findOne({ email: req.body.email, groups: { $nin: req.body.id } }).then(
+    (data) => {
+      if (!data) {
+        res.json({
+          result: false,
+          error: "Utilisateur dans le groupe ou inexistant",
+        });
+        return;
+      }
+      User.updateOne(
+        { email: req.body.email },
+        { $push: { groups: req.body.id } }
+      ).then(() => res.json({ result: true }));
     }
-    User.updateOne(
-      { email: req.body.email },
-      { $push: { groups: req.body.id } }
-    ).then(() => res.json({ result: true }));
-  });
+  );
 });
 
 //route pour quitter un group
@@ -204,7 +152,15 @@ router.put("/exitGroup/:token", (req, res) => {
     if (data) {
       User.updateOne({ token }, { $pull: { groups: req.body.id } }).then(
         (data) => {
-          res.json({ result: true });
+          User.findOne({ groups: req.body.id }).then((data) => {
+            if (data) {
+              res.json({ result: true, info: "Groupe quitté" });
+            } else {
+              Group.deleteOne({ _id: req.body.id  }).then(() =>
+                res.json({ result: true, info: "Groupe supprimé" })
+              );
+            }
+          });
         }
       );
     }
@@ -221,7 +177,6 @@ router.put("/changePassword/:token", (req, res) => {
 
   User.findOne({ token }).then((data) => {
     if (data && bcrypt.compareSync(req.body.oldPassword, data.password)) {
-
       const hash = bcrypt.hashSync(req.body.newPassword, 10);
       User.updateOne({ token }, { password: hash }).then((dataUpdate) =>
         res.json({ result: true })
@@ -233,13 +188,22 @@ router.put("/changePassword/:token", (req, res) => {
 });
 
 router.put("/groups/:token", (req, res) => {
+  if (!checkBody(req.body, ["group"])) {
+    res.json({ result: false, error: "Champs manquants à remplir" });
+    return;
+  }
   const { token } = req.params;
   const { groups } = req.body;
-  User.updateOne({ token }, { groups }).then(() => res.json({ return: true }))
-})
+  User.updateOne({ token }, { groups }).then(() => res.json({ return: true }));
+});
 
 //route qui permet de supprimer un utilisateur via son email
 router.delete("/", (req, res) => {
+  if (!checkBody(req.body, ["email"])) {
+    res.json({ result: false, error: "Champs manquants à remplir" });
+    return;
+  }
+
   User.deleteOne({ email: req.body.email }).then((data) => {
     if (data.deletedCount > 0) {
       res.json({ result: true });
@@ -299,84 +263,65 @@ router.put("/changeEmail/:token", (req, res) => {
   });
 });
 
-router.delete('/', (req, res) => {
-
-  User.deleteOne({ email: req.body.email })
-    .then(data => {
-      console.log(data)
-      if (data.deletedCount > 0) {
-        res.json({ result: true })
-      } else {
-        res.json({ result: false, error: 'No user found' })
-      }
-    }
-
-    )
-})
-
 //route qui permet d'afficher la liste des réservations d'un utilisateur
-
-router.get('/reservations/:token', (req, res) => {
+router.get("/reservations/:token", (req, res) => {
   const { token } = req.params;
-console
-if(token){
-  User.findOne({ token })
-  .populate('reservations.foreignKey')
-    .then(data => {
-      console.log('data', data)
 
-      if (data) {
-        res.json({ result: true, reservations: data.reservations })
-      } else {
-        res.json({ result: false, error: 'No user found' })
-      }
-    }
-    )
+  if (token) {
+    User.findOne({ token })
+      .populate("reservations.foreignKey")
+      .then((data) => {
+        console.log("data", data);
+
+        if (data) {
+          res.json({ result: true, reservations: data.reservations });
+        } else {
+          res.json({ result: false, error: "No user found" });
+        }
+      });
   }
-})
+});
 
 // route qui permet de créer une réservation
-
-router.put('/reservations/:token', (req, res) => {
+router.put("/reservations/:token", (req, res) => {
   const { token } = req.params;
-  const newReservation = { foreignKey: req.body.partnerId, date: req.body.date, group: req.body.group };
-console.log(newReservation)
-  User.findOne({ token })
-    .then(data => {
-      console.log('data', data)
-      if (data) {
-        data.reservations.push(newReservation); // Ajoute la réservation dans le tableau du sous document reservations
-        data.save()
-          .then(updatedUser => {
-            res.json({ result: true, data: updatedUser.reservations });
-          })
-      } else {
-        res.json({ result: false, error: 'No user found' })
-      }
+  if (!checkBody(req.body, ["partnerId", "date", "group"])) {
+    res.json({ result: false, error: "Champs manquants à remplir" });
+    return;
+  }
+  const newReservation = {
+    foreignKey: req.body.partnerId,
+    date: req.body.date,
+    group: req.body.group,
+  };
+  User.findOne({ token }).then((data) => {
+    console.log("data", data);
+    if (data) {
+      data.reservations.push(newReservation); // Ajoute la réservation dans le tableau du sous document reservations
+      data.save().then((updatedUser) => {
+        res.json({ result: true, data: updatedUser.reservations });
+      });
+    } else {
+      res.json({ result: false, error: "No user found" });
     }
-    )
-
-})
+  });
+});
 
 // route qui permet de supprimer une réservation
 
-router.delete('/reservations/:token/:bookingId', (req, res) => {
-  const {token, bookingId} = (req.params);
-console.log(req.params)
+router.delete("/reservations/:token/:bookingId", (req, res) => {
+  const { token, bookingId } = req.params;
 
-  User.findOne({token})
-  .then(data => {
+  User.findOne({ token }).then((data) => {
     if (data) {
+      data.reservations.pull(bookingId); // Retire la réservation du tableau par son id
 
-    data.reservations.pull(bookingId) // Retire la réservation du tableau par son id
-           
-      data.save()
-        .then(updatedUser => {
-          res.json({ result: true, data: updatedUser.reservations });
-        })
+      data.save().then((updatedUser) => {
+        res.json({ result: true, data: updatedUser.reservations });
+      });
     } else {
-      res.json({ result: false, error: 'No user found' })
+      res.json({ result: false, error: "No user found" });
     }
-  })
-})
+  });
+});
 module.exports = router;
